Rational Ray implemented - Still to be investigated

+++ arrangements/external/Cinolib/include/cinolib/Poisson_sampling.cpp
@@ -37,6 +37,7 @@
 #include <cinolib/random_generator.h>
 #include <cinolib/serialize_index.h>
 #include <cinolib/min_max_inf.h>
+#include <array>
 
 namespace cinolib
 {
@@ -174,7 +175,7 @@ void Poisson_sampling(const double          radius,
 
         if(found_sample)
         {
-            uint id=samples.size(); // the index of the new sample
+            uint id=uint(samples.size()); // the index of the new sample
             samples.push_back(x);
             active_list.push_back(id);
             index=serialize_nD_index<Dim,Point>(dim_extent, (x-min)/step);
